fix: rename apidocs to a unique name to avoid weird bug pointing to sdap swagger
